restructure http mod
